Add seeds to the init() call
Add seeds to the init() call so that when the toggles are initialized
they can be initialized with a default state other than false for the
time period prior to fetching them from the remote.

In order to accomplish this I had to no longer use the GBSDKBuilderApp
and drop down to using the GrowthBookSDK class similar to the test
instantiation.

[changelog]
added: seeds to the init() method

ps-id: cf51a770-7bc5-42d0-9802-b7a6c64bd463