Refine local overrides via JSON to use assets
Refine local overrides via JSON to use assets so that we support the
primary use case of working in a flutter application.

The approach was move the overrides to a standard Map structure and then
split out the logic for fetching those overrides to a separate function
to load from the rootBundle assets. In theory this will allow us in the
future to add other mechanisms for fetching overrides if we want without
having to change the core library.

ps-id: b5d60d2f-c4e5-41b9-b3ac-aaad1fe77b5e