Change loadOverridesFromAssets to return none on error
Change loadOverridesFromAssets to return none on error. I made this
switch specifically because on Flutter Web it turns out that the
rootBundle.loadString() doesn't error with a 404 when an asset is
missing. Instead it seems to instead return the index.html. This sucks
because if we were able to get the 404 then we would know it was missing
in the web case and handle that scenario explicitly. Because we don't
have this the best we can do generically hondle it.

This simply means that we have to be aware as local devs to realize when
we flip a toggle and it isn't working as expected and that it might be
due to invalid json format, etc.

[changelog]
fixed: bug where flutter web loadOverridesFromAsets didn't work when missing the asset file

ps-id: a0238725-3e00-4703-88d1-7742c48c8527