Add support for feature types other than boolean
The intention is to support the other data types a feature can be.
These include String, int, or a json data object.

The reason for this is to allow for more flexibility in the types of
features we support. Specifically, I wanted to support String and int
types so we could use this to manage more configuration type things rather
than strictly using this for boolean feature toggles.

[changelog]
added: Support for non-boolean features

ps-id: 9cf32380-e5c7-45d5-a598-bf21b1dc3af2