[UE-2] Add rules parameter to test client
The reason for this change is to be able to test the attributes
feature using our test client. This allows us to eventually pass
app given attributes to the SDK and compare them to rules set
in the SDK to turn a feature on or off. Using this feature, users
of a package can select specific user ids to canary test the app
before releasing to the public. Previously, the test client did not
have the ability to have rules associated with a feature, which
is present in the Growthbook SDK client.

To achieve this, an optional rules parameter has been added
to the test client class and the rules key has been added to
the mapping function on the class.

The optional rules parameter has also been added to the test init
function in the wrapper so that it can pass the necessary rules
to the client.

For example, if the following rules argument is passed to the test
init function:
```[
    {
      'condition': {
        'version': {'$gt': '1.0.0'}
      },
      'force': true
    }
];```
The test client will be able to check for a 'version' attribute and
toggle the feature on or off based on the attribute's value (will
toggle on if the value is greater than 1.0.0.

[changelog]
added: rule parameter to the test client and the test init method

ps-id: E724A507-B0ED-43CA-823D-DFC73BB4899E