[UE-2] Add attributes parameter to init
The intention for this change is to be able to add attributes
from the app (userId, version number, etc) to GrowthBook
and either show or hide a feature based on that feature's
rules. The reason for adding attributes is to allow for canary
testing or limiting features to specific versions of an app.
 This happens on initialization as that is the way to
send attributes to the SDK in flutter

To achieve this, an optional attributes parameter has been added to
the init and test init classes. The attributes are saved to a value on the
class and used in the create client methods when creating the context.
Tests for this functionality use a version number attribute to test
against rules that dictate the app should be a certain version or higher
to use the mock feature

[changelog]
added: attribute parameter to the init methods

ps-id: 6DDF347D-6BAB-4D7B-A31B-AD0B0D421338