[UE-3] Change apiKey to accept full url
The intention for this change is for users of this package to add
their auth details in a more intuitive way. It has been proven that
users instinctively use the full cdn url with the api key in the apiKey,
resulting in an error.

The argument is renamed to apiKeyUrl to be clear about what should
be inputted as an argument. The url is put through a method that removes
the url portion of the string, leaving only the api key to be used in the
`GBContext`. The README is also updated to reflect
these changes and instructs to use the full url found on the GrowthBook
dashboard.

[changelog]
changed: required key for UptechGrowthBookWrapper

ps-id: A7F6E07A-23EB-4EF3-B45E-DA8567C78614