[UE-15] Add set attribute method
The intention for this change is to give users the ability to add
attributes after the wrapper has been initialized. This is helpful
in insteances where required attributes cannot be acquired until
long after an app is initialized. For instance, with canary
testing a feature will only be available to select users.
The user's id cannot be determined until after they have logged
in.

To achieve this, a new public method has been added to the wrapper
that takes in an attributes parameter. If the client already has
attributes (ie they were added during init), the new attributes
are added alongside the old attributes. Otherwise, the new
attributes are assigned to the attributes key in the client's
context.

[changelog]
added: public set attribute method

ps-id: 8E6ED4DD-BB8E-4C27-85E3-4D1F85DF0667