[UE-22] Document setting attributes at initialization
The reason for this change is to give users of this
package instruction on how to add attributes on
init, as well as an explanation on why they may want
to do so. The example shows how they might add a
version attribute on init.

Please note that there are no instructions on adding
the versioning rule in Growthbook or how to get an
app's version number, as this is out of scope for the
package.

[changelog]
added: 'Set attribute on init' section in README

ps-id: D6F4ADEF-5AEC-42AF-9F63-587DEBD716E4