[UE-20] Document non-boolean toggle state
The reason for this change is to give users of this package
instruction on how to use the value method. The
example shows how they might get a feature's value (in this
case an int) and use it to determine some business logic.

[changelog]
added: 'Evaluate Feature value' section in README

ps-id: 1A743160-40E2-4E30-976C-CA4D31C1B55B