Update init to take features parameter: Map<String, GBFeature>
Recently the growthbook_sdk_flutter package was updated. This change
removed the ability to "seed" the feature values which are used while
the network request is happening or as fallback if the request fails.

Alongside this work, I have updated the Uptech fork of
growthbook_sdk_flutter to allow passing in "seeded" features and
values at init, instead of after. This commits points the dependency
at the Uptech fork, which is still in review.

Add features param to init constructor.
Add Deprecated tag to seeds so we know, while it is still possible to
use that parameter, we should be using features.
Add comments explaining why we have both params and what changes in
the growthbook_sdk_flutter package inspired these changes.
If a Map is passed in to the seeds param, we convert them to the new
feature Map<String, GBFeature> and pass that map along.
Update the tests to use GBFeature for map values.

<!-- ps-id: 9fb1da74-fa5f-4ba5-a05e-ddebf7f970b6 -->